+++ R/server.R
@@ -1,12 +1,16 @@
 #' Create a SMTP server object.
 #'
+#' @name server
+#'
 #' @param host DNS name or IP address of the SMTP server.
 #' @param port Port that the SMTP server is listening on.
 #' @param username Username for SMTP server.
-#' @param password Password for SMTP server.
+#' @param password Password for SMTP server or API key.
 #' @param insecure Whether to ignore SSL issues.
 #' @param reuse Whether the connection to the SMTP server should be left open for reuse.
 #' @param helo The HELO domain name of the sending host. If left as \code{NA} then will use local host name.
+#' @param protocol Which protocol (SMTP or SMTPS) to use for communicating with
+#'   the server. Default will choose appropriate protocol based on port.
 #' @param max_times Maximum number of times to retry.
 #' @param pause_base Base delay (in seconds) for exponential backoff. See \link[purrr]{rate_backoff}.
 #' @param ... Additional curl options. See \code{curl::curl_options()} for a list of supported options.
@@ -16,10 +20,10 @@
 #' @examples
 #' library(magrittr)
 #'
-#' # Set parameters for SMTP server (with username and password)
+#' # Set parameters for SMTP server (with username and password).
 #' smtp <- server(
 #'   host = "smtp.gmail.com",
-#'   port = 465,
+#'   port = 587,
 #'   username = "bob@gmail.com",
 #'   password = "bd40ef6d4a9413de9c1318a65cbae5d7"
 #' )
@@ -61,6 +65,7 @@ server <- function(
   insecure = FALSE,
   reuse = TRUE,
   helo = NA,
+  protocol = NA,
   pause_base = 1,
   max_times = 5,
   ...) {
@@ -100,12 +105,10 @@ server <- function(
       use_ssl = 0
     }
 
-    helo <- ifelse(is.na(helo), "", helo)
-
-    smtp_server <- sprintf("%s:%d/%s", host, port, helo)
+    smtp_server <- smtp_url(host, port, protocol, helo)
     #
     if (verbose) {
-      cat("Sending email to ", smtp_server, ".\n", file = stderr(), sep = "")
+      log_debug("Sending email to ", smtp_server, ".")
     }
 
     # Create an insistent version of send_mail().
@@ -145,3 +148,100 @@ server <- function(
     invisible(result)
   }
 }
+
+#' @rdname server
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # Set parameters for Gmail SMTP server. The host and port are implicit.
+#' smtp <- gmail(
+#'   username = "bob@gmail.com",
+#'   password = "bd40ef6d4a9413de9c1318a65cbae5d7"
+#' )
+gmail <- function(
+  username,
+  password,
+  ...) {
+  # Retrieve function.
+  fcall <- match.call(expand.dots = TRUE)
+  # Substitute server() as function to call.
+  fcall[[1]] <- server
+  # Fill in host and port arguments.
+  fcall$host = "smtp.gmail.com"
+  fcall$port = ifelse(is.null(fcall$port), 587, fcall$port)
+  # Now run the function call.
+  eval(fcall, parent.frame())
+}
+
+#' @rdname server
+#'
+#' @section Sendgrid:
+#'
+#' To use SendGrid you'll need to first \href{https://docs.sendgrid.com/for-developers/sending-email/integrating-with-the-smtp-api}{create an API key}.
+#' Then use the API key as the password.
+#'
+#' SendGrid will accept messages on ports 25, 587 and 2525 (using SMTP) as well
+#' as 465 (using SMTPS).
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # Set API key for SendGrid SMTP server.
+#' smtp <- sendgrid(
+#'   password = "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"
+#' )
+sendgrid <- function(
+  password,
+  ...) {
+  fcall <- match.call(expand.dots = TRUE)
+
+  fcall[[1]] <- server
+  fcall$host = "smtp.sendgrid.net"
+  fcall$port = ifelse(is.null(fcall$port), 587, fcall$port)
+  fcall$username = "apikey"
+
+  eval(fcall, parent.frame())
+}
+
+#' @rdname server
+#'
+#' @section Mailgun:
+#'
+#' To use Mailgun you'll need to first register a sender domain. This will then
+#' be assigned a username and password.
+#'
+#' Mailgun will accept messages on ports 25 and 587 (using SMTP) as well as 465
+#' (using SMTPS).
+#'
+#' @export
+#'
+#' @examples
+#'
+#' # Set username and password for Mailgun SMTP server.
+#' smtp <- mailgun(
+#'   username = "postmaster@sandbox9ptce35fdf0b31338dec4284eb7aaa59.mailgun.org",
+#'   password = "44d072e7g2b5f3bf23b2b642da0fe3a7-2ac825a1-a5be680a"
+#' )
+mailgun <- function(
+  username,
+  password,
+  ...) {
+  fcall <- match.call(expand.dots = TRUE)
+
+  fcall[[1]] <- server
+  fcall$host = "smtp.mailgun.org"
+  fcall$port = ifelse(is.null(fcall$port), 587, fcall$port)
+
+  eval(fcall, parent.frame())
+}
+
+# ./swaks --auth \
+# --server  \
+# --au postmaster@YOUR_DOMAIN_NAME \
+# --ap 3kh9umujora5 \
+# --to bar@example.com \
+# --h-Subject: "Hello" \
+# --body 'Testing some Mailgun awesomness!'